+++ Changelog.md
@@ -6,6 +6,7 @@ Bugfixes:
  * Type Checker: Dissallow mappings with data locations other than 'storage'
  * Type Checker: Fix internal error when a struct array index does not fit into a uint256.
  * Type system: Properly report packed encoded size for arrays and structs (mostly unused until now).
+ * Commandline interface: Allow yul optimizer only for strict assembly.
 
 
 Language Features: